# src/services/connector-service/openapi.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "_IssueWire": {
        "additionalProperties": false,
        "description": "Wire shape of the ``issue`` request body.\n\n``workspaceId`` is intentionally absent \u2014 it comes from the\ncall-context header so the sidecar cannot forge a lease across\nworkspaces.",
        "properties": {
          "attempt": {
            "minimum": 1.0,
            "title": "Attempt",
            "type": "integer"
          },
          "capability": {
            "minLength": 1,
            "title": "Capability",
            "type": "string"
          },
          "connectorInstanceId": {
            "minLength": 1,
            "title": "Connectorinstanceid",
            "type": "string"
          },
          "instanceTtlSec": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instancettlsec"
          },
          "requestedTtlSec": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requestedttlsec"
          },
          "runId": {
            "minLength": 1,
            "title": "Runid",
            "type": "string"
          },
          "slot": {
            "minLength": 1,
            "title": "Slot",
            "type": "string"
          },
          "stepDeadline": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stepdeadline"
          },
          "stepId": {
            "minLength": 1,
            "title": "Stepid",
            "type": "string"
          },
          "tokenType": {
            "minLength": 1,
            "title": "Tokentype",
            "type": "string"
          },
          "typeMaxTtlSec": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Typemaxttlsec"
          }
        },
        "required": [
          "runId",
          "stepId",
          "attempt",
          "slot",
          "capability",
          "connectorInstanceId",
          "tokenType"
        ],
        "title": "_IssueWire",
        "type": "object"
      },
...
      "_ReleaseWire": {
        "additionalProperties": false,
        "description": "Wire shape of the ``release`` request body.",
        "properties": {
          "leaseId": {
            "minLength": 1,
            "title": "Leaseid",
            "type": "string"
          }
        },
        "required": [
          "leaseId"
        ],
        "title": "_ReleaseWire",
        "type": "object"
      },
      "_RevokeRequest": {
        "additionalProperties": false,
        "description": "Body for ``POST /v1/workspaces/{ws}/leases/{leaseId}:revoke``.",
        "properties": {
          "reason": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "_RevokeRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Custos Connector Service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/internal/v1/leases:issue": {
      "post": {
        "description": "Mint a new lease for the calling sidecar.\n\nThe ``workspaceId`` is taken from the call-context (not the body).\nOn success returns the full lease envelope; on ``LeaseError`` the\nresponse carries the ``{\"error\": {\"code\", \"detail\"}}`` envelope and\nthe HTTP status :func:`http_status_for` selects.",
        "operationId": "issue_lease_internal_v1_leases_issue_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/_IssueWire"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Issue Lease",
        "tags": [
          "lease"
        ]
      }
    },
    "/internal/v1/leases:refresh": {
      "post": {
        "description": "Extend the expiry of an existing lease without changing its id.",
        "operationId": "refresh_lease_internal_v1_leases_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/_RefreshWire"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh Lease",
        "tags": [
          "lease"
        ]
      }
    },
    "/internal/v1/leases:release": {
      "post": {
        "description": "Mark a lease as released. Always 204 (idempotent, best-effort).\n\nThe design contract: release is best-effort and clients should not\ndistinguish \"lease found and released\" from \"lease unknown / already\nreleased\". Either way the activity's hold on the upstream credential\nis over.",
        "operationId": "release_lease_internal_v1_leases_release_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/_ReleaseWire"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Release Lease",
        "tags": [
          "lease"
        ]
      }
    },
    "/internal/v1/leases:revoke": {
      "post": {
        "description": "Revoke a batch of leases. Returns a per-lease ack list (CONN-IMPL-020).\n\nBody shape: ``{leaseIds: [...], reason: \"...\"}``.\n\nResponse shape: ``{results: [{leaseId, status}, ...]}`` where\n``status`` is one of ``revoked`` / ``already-revoked`` /\n``already-expired`` / ``not-found`` (see\n:class:`~custos_connector.lease.service.RevokeOutcomeStatus`).\n\nThe endpoint always returns 200; per-lease failures are not HTTP\nerrors. This matches the sidecar control-channel contract \u2014 the\ncaller is interested in the per-lease outcome, not a single\npass/fail for the whole batch. Duplicate ids in the request body\nproduce one ack each, in input order; the second ack for a\nduplicate sees ``already-revoked`` because the first call moved\nthe lease into the revoked state.",
        "operationId": "revoke_leases_internal_v1_leases_revoke_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/_RevokeWire"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke Leases",
        "tags": [
          "lease"
        ]
      }
    },
...
    "/v1/workspaces/{ws}/audit/leases": {
      "get": {
        "description": "List lease audit events for a workspace, newest-first.\n\nWhen ``eventType`` is omitted, the route fetches without an\nSPL-level event filter and post-filters to events whose\n``event_type`` is one of the known lease event types in\n:data:`_LEASE_EVENT_TYPES`. Callers narrowing to a single lease\nevent type SHOULD pass ``eventType`` so the SPL surface does the\nfilter and pagination stays dense.",
        "operationId": "list_lease_audit_events_v1_workspaces__ws__audit_leases_get",
        "parameters": [
          {
            "in": "path",
            "name": "ws",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Ws",
              "type": "string"
            }
          },
          {
            "description": "Optional single event-type filter. Must be one of: lease.issued, lease.refreshed, lease.released, lease.expired, lease.revoke-requested, lease.revoked, lease.denied. Omit to query all lease events.",
            "in": "query",
            "name": "eventType",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional single event-type filter. Must be one of: lease.issued, lease.refreshed, lease.released, lease.expired, lease.revoke-requested, lease.revoked, lease.denied. Omit to query all lease events.",
              "title": "Eventtype"
            }
          },
          {
            "in": "query",
            "name": "actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Actor"
            }
          },
          {
            "in": "query",
            "name": "occurredAfter",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Occurredafter"
            }
          },
          {
            "in": "query",
            "name": "occurredBefore",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Occurredbefore"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 500,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Query the audit log for lease-related events.",
        "tags": [
          "audit"
        ]
      }
    },
...
    "/v1/workspaces/{ws}/connectors/{instance_id}/leases": {
      "get": {
        "operationId": "list_leases_for_instance_v1_workspaces__ws__connectors__instance_id__leases_get",
        "parameters": [
          {
            "in": "path",
            "name": "ws",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Ws",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "instance_id",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Instance Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 500,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List active leases for a connector instance.",
        "tags": [
          "lease-admin"
        ]
      }
    },
    "/v1/workspaces/{ws}/connectors/{instance_id}/leases:revoke-all": {
      "post": {
        "operationId": "revoke_all_for_instance_v1_workspaces__ws__connectors__instance_id__leases_revoke_all_post",
        "parameters": [
          {
            "in": "path",
            "name": "ws",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Ws",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "instance_id",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Instance Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/_RevokeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke every active lease for a connector instance.",
        "tags": [
          "lease-admin"
        ]
      }
    },
...
    "/v1/workspaces/{ws}/leases/{lease_id}:revoke": {
      "post": {
        "description": "Revoke one lease.\n\nReturns 200 with ``{leaseId, status}`` when the lease was found\n(``status`` is one of ``revoked`` / ``already-revoked`` /\n``already-expired``); returns 404 with the same envelope when\nthe lease does not exist in this workspace.",
        "operationId": "revoke_single_lease_v1_workspaces__ws__leases__lease_id__revoke_post",
        "parameters": [
          {
            "in": "path",
            "name": "ws",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Ws",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "lease_id",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Lease Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/_RevokeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke a single lease.",
        "tags": [
          "lease-admin"
        ]
      }
    },
    "/v1/workspaces/{ws}/runs/{run_id}/leases": {
      "get": {
        "operationId": "list_leases_for_run_v1_workspaces__ws__runs__run_id__leases_get",
        "parameters": [
          {
            "in": "path",
            "name": "ws",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Ws",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 500,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List active leases for a run.",
        "tags": [
          "lease-admin"
        ]
      }
    },
    "/v1/workspaces/{ws}/runs/{run_id}/leases:revoke-all": {
      "post": {
        "operationId": "revoke_all_for_run_v1_workspaces__ws__runs__run_id__leases_revoke_all_post",
        "parameters": [
          {
            "in": "path",
            "name": "ws",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Ws",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/_RevokeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke every active lease for a run.",
        "tags": [
          "lease-admin"
        ]
      }
    }
  }
}